refactor(cli): use a unified type across all CLI generators
