--- lib/cli/src/GeneratorOptions.ts
@@ -1,5 +1,0 @@
-import { StoryFormat } from './project_types';
-
-export type GeneratorOptions = {
-  storyFormat: StoryFormat;
-};
