+++ lib/cli/src/generators/HTML/index.ts
@@ -7,10 +7,9 @@ import {
   copyTemplate,
 } from '../../helpers';
 import { StoryFormat } from '../../project_types';
-import { NpmOptions } from '../../NpmOptions';
-import { GeneratorOptions } from '../../GeneratorOptions';
+import { Generator } from '../Generator';
 
-export default async (npmOptions: NpmOptions, { storyFormat }: GeneratorOptions) => {
+const generator: Generator = async (npmOptions, { storyFormat }) => {
   const packages = ['@storybook/html', '@storybook/addon-actions', '@storybook/addon-links'];
 
   const versionedPackages = await getVersionedPackages(npmOptions, ...packages);
@@ -35,3 +34,5 @@ export default async (npmOptions: NpmOptions, { storyFormat }: GeneratorOptions)
 
   installDependencies({ ...npmOptions, packageJson }, [...versionedPackages, ...babelDependencies]);
 };
+
+export default generator;
