+++ lib/cli/src/generators/WEB-COMPONENTS/index.ts
@@ -8,10 +8,9 @@ import {
   installDependencies,
 } from '../../helpers';
 import { StoryFormat } from '../../project_types';
-import { NpmOptions } from '../../NpmOptions';
-import { GeneratorOptions } from '../../GeneratorOptions';
+import { Generator } from '../Generator';
 
-export default async (npmOptions: NpmOptions, { storyFormat }: GeneratorOptions) => {
+const generator: Generator = async (npmOptions, { storyFormat }) => {
   const packages = [
     '@storybook/web-components',
     '@storybook/addon-actions',
@@ -42,3 +41,5 @@ export default async (npmOptions: NpmOptions, { storyFormat }: GeneratorOptions)
 
   installDependencies({ ...npmOptions, packageJson }, [...versionedPackages, ...babelDependencies]);
 };
+
+export default generator;
