+++ lib/cli/src/generators/WEBPACK_REACT/index.ts
@@ -7,10 +7,9 @@ import {
   copyTemplate,
 } from '../../helpers';
 import { StoryFormat } from '../../project_types';
-import { NpmOptions } from '../../NpmOptions';
-import { GeneratorOptions } from '../../GeneratorOptions';
+import { Generator } from '../Generator';
 
-export default async (npmOptions: NpmOptions, { storyFormat }: GeneratorOptions) => {
+const generator: Generator = async (npmOptions, { storyFormat }) => {
   const packages = [
     '@storybook/react',
     '@storybook/addon-actions',
@@ -39,3 +38,5 @@ export default async (npmOptions: NpmOptions, { storyFormat }: GeneratorOptions)
 
   installDependencies({ ...npmOptions, packageJson }, [...versionedPackages, ...babelDependencies]);
 };
+
+export default generator;
